requirements.txt
drawsvg==2.4.0
Pillow==11.0.0
pybrl==0.1.0
